requirements.txt
beautifulsoup4==4.11.1
python-dotenv==0.20.0
python-telegram-bot==20.0a4
requests==2.28.1
